Allow generation of sortkeys via templates rather than a special 'generate_sortkey()' function.
This required a number of changes in several places of bibulous.py, largely because attempting this goal uncovered a bug with the implicit loops --- if all variables inside the loop structure are undefined, then it returns the loop ellipsis bracketed by the '???' undefstr replacements. Not good. A better response is to return just 'None', as the code now does.