Modified bibulous.py's handling of citation keys that are missing from the database.
Where before bibulous did not even create an entrykey for this case, now it creates a default entrykey and fills it with minimal info plus an error message. This allows all of the code to treat it normally, simplifies the code logic a little, and allows the output to insert the entry error message at its proper place in the citation list.

Also modified test1 in the testing suite to accommodate the change.